requirements.txt
numpy
pandas
matplotlib
seaborn
pillow
scikit-learn
tensorflow-macos
tensorflow-metal
